# DriveSense-nest-api/docs/api.yaml
openapi: 3.0.0
info:
  title: DriveSense API
  description: API para gestionar restricciones de tráfico (pico y placa) en ciudades
  version: 1.0.0
  contact:
    name: DriveSense Team
servers:
  - url: http://localhost:4000
    description: Local Development Server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from login
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: John Doe
        email:
          type: string
          format: email
          example: john@example.com
        role:
          type: string
          enum: [admin, user]
          example: user
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateUserDto:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          minLength: 1
          example: John Doe
        email:
          type: string
          format: email
          example: john@example.com
        password:
          type: string
          minLength: 8
          example: SecurePass123
        role:
          type: string
          enum: [admin, user]
          example: user
...
    LoginDto:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: john@example.com
        password:
          type: string
          minLength: 8
          maxLength: 32
          example: SecurePass123
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        user:
          $ref: '#/components/schemas/User'

    # City Schemas
    City:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Bogotá
        description:
          type: string
          example: Capital de Colombia
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateCityDto:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 3
          example: Bogotá
        description:
          type: string
          example: Capital de Colombia
        isActive:
          type: boolean
          example: true
    UpdateCityDto:
      type: object
      properties:
        name:
          type: string
          minLength: 3
          example: Medellín
        description:
          type: string
          example: Ciudad de la Eterna Primavera
        isActive:
          type: boolean
          example: true

    # Vehicle Schemas
...
    Rule:
      type: object
      properties:
        id:
          type: integer
          example: 1
        dayOfWeek:
          type: string
          example: Lunes
        startTime:
          type: string
          example: "06:00"
        endTime:
          type: string
          example: "08:30"
        restrictedDigits:
          type: array
          items:
            type: string
          example: ["1", "2"]
        isActive:
          type: boolean
          example: true
        city:
          $ref: '#/components/schemas/City'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateRuleDto:
      type: object
      required:
        - dayOfWeek
        - startTime
        - endTime
        - restrictedDigits
        - city
      properties:
        dayOfWeek:
          type: string
          example: Lunes
        startTime:
          type: string
          pattern: '^\d{2}:\d{2}$'
          example: "06:00"
        endTime:
          type: string
          pattern: '^\d{2}:\d{2}$'
          example: "08:30"
        restrictedDigits:
          type: array
          items:
            type: string
          example: ["1", "2"]
        isActive:
          type: boolean
          example: true
        city:
          type: integer
          example: 1
    UpdateRuleDto:
      type: object
      properties:
        dayOfWeek:
          type: string
          example: Martes
        startTime:
          type: string
          pattern: '^\d{2}:\d{2}$'
          example: "07:00"
        endTime:
          type: string
          pattern: '^\d{2}:\d{2}$'
          example: "09:30"
        restrictedDigits:
          type: array
          items:
            type: string
          example: ["3", "4"]
        isActive:
          type: boolean
          example: true

    # Error Responses
    UnauthorizedResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 401
        message:
          type: string
          example: Unauthorized
        error:
          type: string
          example: Unauthorized
    ForbiddenResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 403
        message:
          type: string
          example: Forbidden resource
        error:
          type: string
          example: Forbidden
    NotFoundResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 404
        message:
          type: string
          example: Resource not found
        error:
          type: string
          example: Not Found
    BadRequestResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: array
          items:
            type: string
          example: ["Validation error"]
        error:
          type: string
          example: Bad Request
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Creates a new user account in the system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        201:
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        400:
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestResponse'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: Authenticate user with email and password to get JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        200:
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        400:
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestResponse'
  /auth/profile:
    get:
      tags:
        - Authentication
      summary: Get user profile
      description: Retrieve the authenticated user's profile information
      security:
        - BearerAuth: []
      responses:
        200:
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        401:
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'

  # Users Endpoints
...
  /city:
    post:
      tags:
        - Cities
      summary: Create a new city
      description: Create a new city with traffic restriction rules (Admin only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCityDto'
      responses:
        201:
          description: City created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/City'
        400:
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        403:
          description: Forbidden - Admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'

    get:
      tags:
        - Cities
      summary: Get all cities
      description: List all cities in the system
      security:
        - BearerAuth: []
      responses:
        200:
          description: Cities retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/City'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
  /city/{id}:
    get:
      tags:
        - Cities
      summary: Get city by ID
      description: Retrieve a specific city by its ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        200:
          description: City retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/City'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        404:
          description: City not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundResponse'

    patch:
      tags:
        - Cities
      summary: Update city
      description: Update city information (Admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCityDto'
      responses:
        200:
          description: City updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/City'
        400:
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        403:
          description: Forbidden - Admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'
        404:
          description: City not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundResponse'

    delete:
      tags:
        - Cities
      summary: Delete city
      description: Remove a city from the system (Admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        200:
          description: City deleted successfully
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        403:
          description: Forbidden - Admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'
        404:
          description: City not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundResponse'

  # Vehicles Endpoints
...
  /rules:
    post:
      tags:
        - Rules
      summary: Create a new rule
      description: Create a new traffic restriction rule for a city (Admin only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRuleDto'
      responses:
        201:
          description: Rule created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rule'
        400:
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        403:
          description: Forbidden - Admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'

    get:
      tags:
        - Rules
      summary: Get all rules
      description: List all traffic restriction rules (Admin only)
      security:
        - BearerAuth: []
      responses:
        200:
          description: Rules retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Rule'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        403:
          description: Forbidden - Admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'
  /rules/{id}:
    get:
      tags:
        - Rules
      summary: Get rule by ID
      description: Retrieve a specific traffic restriction rule by its ID (Admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        200:
          description: Rule retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rule'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        403:
          description: Forbidden - Admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'
        404:
          description: Rule not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundResponse'

    patch:
      tags:
        - Rules
      summary: Update rule
      description: Update a traffic restriction rule (Admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRuleDto'
      responses:
        200:
          description: Rule updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rule'
        400:
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        403:
          description: Forbidden - Admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'
        404:
          description: Rule not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundResponse'

    delete:
      tags:
        - Rules
      summary: Delete rule
      description: Remove a traffic restriction rule (Admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        200:
          description: Rule deleted successfully
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        403:
          description: Forbidden - Admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'
        404:
          description: Rule not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundResponse'
...